fix: refactor convert functions to utils
